tranformando a orientado a objetos
